+++ src/components/cards/ContactCard.js
@@ -1,0 +1,9 @@
+import React, { useState, useEffect } from "react";
+
+export default function ContactCard() {
+    return (
+        <>
+            <h1>hehe</h1>
+        </>
+    )
+}
