+++ src/components/cards/NoteCard.js
@@ -1,34 +1,75 @@
 import React, { useState, useEffect } from "react";
-import { Grid, Card, CardHeader, CardContent, Typography, IconButton } from "@material-ui/core";
+import {
+  Container,
+  Grid,
+  Card,
+  CardHeader,
+  CardContent,
+  Typography,
+  IconButton,
+} from "@material-ui/core";
 import { DeleteOutlined } from "@material-ui/icons";
 
-export default function NoteCard(props) {
-  const { note, index, data, setData } = props
-  
+const noteObj = [
+  {
+    title: "Yoshi's birthday bash",
+    details:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book.",
+    category: "reminders",
+    id: 1,
+  },
+  {
+    title: "Complete my ninja training",
+    details:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took.",
+    category: "work",
+    id: 2,
+  },
+  {
+    title: "Order a pizza!",
+    details:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book.\nLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book.",
+    category: "todos",
+    id: 3,
+  },
+];
+
+// card container
+export default function NoteCard() {
+  const [notes, setNotes] = useState(noteObj || []);
+
   const handleDelete = (id) => {
-    if (data.length > 0) {
-      setData(data.filter((card, index) => index !== id))
+    if (notes.length > 0) {
+      setNotes(notes.filter((card, index) => index !== id));
     }
-  }
+  };
 
   return (
     <>
-      <Card elevation={1}>
-        <CardHeader
-          action={
-            <IconButton onClick={() => handleDelete(index)}>
-              <DeleteOutlined />
-            </IconButton>
-          }
-          title={note.title}
-          subheader={note.category}
-        />
-        <CardContent>
-          <Typography variant="body2" color="textSecondary">
-            {note.details}
-          </Typography>
-        </CardContent>
-      </Card>
+      <Container>
+        <Grid container spacing={3}>
+          {notes.map((note, id) => (
+            <Grid item key={id} xs={12} sm={6} md={4}>
+              <Card elevation={1}>
+                <CardHeader
+                  action={
+                    <IconButton onClick={() => handleDelete(id)}>
+                      <DeleteOutlined />
+                    </IconButton>
+                  }
+                  title={note.title}
+                  subheader={note.category}
+                />
+                <CardContent>
+                  <Typography variant="body2" color="textSecondary">
+                    {note.details}
+                  </Typography>
+                </CardContent>
+              </Card>
+            </Grid>
+          ))}
+        </Grid>
+      </Container>
     </>
   );
-};
+}
